+++ R/summary.R
@@ -3,13 +3,13 @@
 #' Displays a tree-like summary of the elements that have been added to an
 #' OmicNavigator study.
 #'
-#' @param object OmicNavigator study object (class \code{oaStudy})
+#' @param object OmicNavigator study object (class \code{onStudy})
 #' @param elements Subset the output to only include specific elements of the
 #'   study, e.g. \code{c("results", "enrichments")}
 #' @param ... Currently unused
 #'
 #' @export
-summary.oaStudy <- function(object, elements = NULL, ...) {
+summary.onStudy <- function(object, elements = NULL, ...) {
   if (!is.null(elements) && is.character(elements)) {
     object <- object[elements]
   }