+++ inst/tinytest/testPrint.R
@@ -1,4 +1,4 @@
-# Test print.oaStudy()
+# Test print.onStudy()
 
 # Setup ------------------------------------------------------------------------
 
@@ -11,7 +11,7 @@ emptyStudy <- createStudy(name = "empty", description = "An empty study")
 testStudy <- OmicNavigator:::testStudy(name = "test", description = "A test study")
 testStudyPlots <- addPlots(testStudy, OmicNavigator:::testPlots())
 
-# Test print.oaStudy() ---------------------------------------------------------
+# Test print.onStudy() ---------------------------------------------------------
 
 expect_stdout(
   print(emptyStudy),