+++ inst/tinytest/testSummary.R
@@ -1,4 +1,4 @@
-# Test summary.oaStudy()
+# Test summary.onStudy()
 
 # Setup ------------------------------------------------------------------------
 
@@ -11,7 +11,7 @@ emptyStudy <- createStudy(name = "empty", description = "An empty study")
 testStudy <- OmicNavigator:::testStudy(name = "test", description = "A test study")
 testStudyPlots <- addPlots(testStudy, OmicNavigator:::testPlots())
 
-# Test summary.oaStudy() -------------------------------------------------------
+# Test summary.onStudy() -------------------------------------------------------
 
 expect_stdout(
   summary(emptyStudy),